Update image availability check in main.py
This change refactors checking for new images in the main.py file, prioritizing the live view. Instead of basing it on whether a client is connected, images are now sent based on the value of the 'enable_live_view' field in the plc_handler's vs_ctrl object. The edge position and frames per second (fps) for both rear and front cameras are also updated within each camera's new image check.